addressing review follow up from pull 3086
exports some more things from aria/grid package and adds a new hook for selection announcements
