--- packages/@react-aria/list/intl/en-US.json
@@ -1,8 +1,0 @@
-{
-  "deselectedItem": "{item} not selected.",
-  "select": "Select",
-  "selectedCount": "{count, plural, =0 {No items selected} one {# item selected} other {# items selected}}.",
-  "selectedAll": "All items selected.",
-  "selectedItem": "{item} selected.",
-  "longPressToSelect": "Long press to enter selection mode."
-}
