--- packages/@react-aria/list/intl/es-ES.json
@@ -1,8 +1,0 @@
-{
-  "deselectedItem": "{item} no seleccionado.",
-  "longPressToSelect": "Mantenga pulsado para abrir el modo de selección.",
-  "select": "Seleccionar",
-  "selectedAll": "Todos los elementos seleccionados.",
-  "selectedCount": "{count, plural, =0 {Ningún elemento seleccionado} one {# elemento seleccionado} other {# elementos seleccionados}}.",
-  "selectedItem": "{item} seleccionado."
-}
