--- packages/@react-aria/list/intl/pt-BR.json
@@ -1,8 +1,0 @@
-{
-  "deselectedItem": "{item} não selecionado.",
-  "longPressToSelect": "Mantenha pressionado para entrar no modo de seleção.",
-  "select": "Selecionar",
-  "selectedAll": "Todos os itens selecionados.",
-  "selectedCount": "{count, plural, =0 {Nenhum item selecionado} one {# item selecionado} other {# itens selecionados}}.",
-  "selectedItem": "{item} selecionado."
-}
