+++ packages/@react-aria/list/src/useListItem.ts
@@ -26,9 +26,7 @@ export interface AriaListItemOptions {
   /** Whether the list row is contained in a virtual scroller. */
   isVirtualized?: boolean,
   /** Whether selection should occur on press up instead of press down. */
-  shouldSelectOnPressUp?: boolean,
-  /** Whether the list item is disabled. */
-  isDisabled?: boolean
+  shouldSelectOnPressUp?: boolean
 }
 
 export interface ListItemAria {
@@ -51,8 +49,7 @@ export function useListItem<T>(props: AriaListItemOptions, state: ListState<T>, 
   let {
     node,
     isVirtualized,
-    shouldSelectOnPressUp,
-    isDisabled
+    shouldSelectOnPressUp
   } = props;
 
   let {direction} = useLocale();
@@ -72,7 +69,6 @@ export function useListItem<T>(props: AriaListItemOptions, state: ListState<T>, 
     isVirtualized,
     shouldSelectOnPressUp,
     onAction: onAction ? () => onAction(node.key) : undefined,
-    isDisabled,
     focus
   });
